refactor ai panel into subcomponents
